+++ packages/hightide-native/src/global-contexts/content-theme/ContentThemeContext.ts
@@ -4,9 +4,12 @@ import {
 } from 'react'
 import type { ColorValue, TextStyle } from 'react-native'
 
+import type { IconStyle } from '../../icons'
+
 export type ContentThemeContextValue = {
-  foregroundColor: ColorValue,
+  foreground: ColorValue,
   textStyle: TextStyle,
+  iconStyle: IconStyle,
 }
 
 export const ContentThemeContext = createContext<ContentThemeContextValue | null>(null)
@@ -15,7 +18,7 @@ export const useContentTheme = (): ContentThemeContextValue => {
   const context = useContext(ContentThemeContext)
   if (!context) {
     throw new Error(
-      'useContentTheme must be used within ContentThemeProvider. Try adding a ContentThemeProvider around your app.'
+      'useContentTheme must be used within ContentThemeRootProvider. Try adding a ContentThemeRootProvider around your app.'
     )
   }
   return context