+++ packages/hightide-native/src/theme/types/components/iconButton.ts
@@ -1,7 +1,4 @@
-import type {
-  TextStyle,
-  ViewStyle
-} from 'react-native'
+import type { ViewStyle } from 'react-native'
 
 import type { ColorPairToken } from '@helpwave/hightide-design/theme-tokens'
 import type {
@@ -25,11 +22,8 @@ export type IconButtonStyle = ViewStyle
 
 export type IconButtonIconStyle = IconStyle
 
-export type IconButtonTextStyle = TextStyle
-
 export type IconButtonThemeResolvers = {
   container: StyleResolverFunction<IconButtonState, IconButtonStyle>,
   stateLayer: StyleResolverFunction<IconButtonState, IconButtonStyle>,
   icon: StyleResolverFunction<IconButtonState, IconButtonIconStyle>,
-  text: StyleResolverFunction<IconButtonState, IconButtonTextStyle>,
 }